refactor(backend): OrderRequestProduct SF Object 정합 — OwnerId polymorphic R-2 + audit User FK + LineChangeType free-form (761)
- OwnerId polymorphic: owner_id (Employee FK) → owner_user_id (User) + owner_group_id (Group) + XOR CHECK
- audit FK: createdBy/lastModifiedBy 타입 Employee → User
- LineChangeType: is_cancelled Boolean 우회 폐기 → line_change_type VARCHAR(10) free-form (SF/Heroku 레거시 운영값 {null, "X"} 분석 결과 반영)
  cancel() / isCancelled() 도메인 메서드로 application 로직 보존, LINE_CHANGE_TYPE_CANCEL = "X" 상수
- 타입·길이 정합 9건: unit/product_code/box_quantity/amount/quantity_boxes length+precision 확장, line_number/quantity_pieces Int→Long, dk_total_count/total_count Double→BigDecimal(18,0)
- V133 마이그레이션 + Spec #623 어노테이션 테스트 + cancel/isCancelled 도메인 테스트

+++ backend/src/test/kotlin/com/otoki/powersales/order/sap/OrderRequestCancelPayloadFactoryTest.kt
@@ -23,8 +23,8 @@ class OrderRequestCancelPayloadFactoryTest {
     fun payloadStructure() {
         val orderRequest = orderRequest()
         val products = listOf(
-            product(id = 101, lineNumber = 10, productCode = "P001", orderRequest = orderRequest),
-            product(id = 102, lineNumber = 20, productCode = "P002", orderRequest = orderRequest),
+            product(id = 101, lineNumber = 10L, productCode = "P001", orderRequest = orderRequest),
+            product(id = 102, lineNumber = 20L, productCode = "P002", orderRequest = orderRequest),
         )
 
         val payload = factory.build(orderRequest, products)
@@ -66,7 +66,7 @@ class OrderRequestCancelPayloadFactoryTest {
 
     private fun product(
         id: Long,
-        lineNumber: Int,
+        lineNumber: Long,
         productCode: String,
         orderRequest: OrderRequest,
     ) = OrderRequestProduct(
